requirements.txt
Django==4.0.6
djangorestframework==3.13.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
drf-yasg==1.20.*
drf-nested-routers==0.93.*
requests==2.28.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
djangorestframework-gis==1.0.*